# swagger.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "info": {
    "title": "MRCS Question Extractor API",
    "description": "AI-powered medical exam question extraction from PDF documents",
    "version": "1.0.0",
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
...
    "/categories": {
      "get": {
        "summary": "Get all active categories",
        "description": "Get all active categories with pagination and filtering",
        "tags": [
          "categories"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDto__type"
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "categories",
          "findAll"
        ],
        "x-samchon-controller": "CategoriesController"
      },
      "post": {
        "summary": "Create a new category",
        "description": "Create a new category",
        "tags": [
          "categories"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "type": {
                      "oneOf": [
                        {
                          "const": "BASIC"
                        },
                        {
                          "const": "CLINICAL"
                        }
                      ]
                    },
                    "questionCount": {
                      "type": "number"
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "name",
                    "id",
                    "displayName",
                    "type",
                    "questionCount",
                    "isActive",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "categories",
          "create"
        ],
        "x-samchon-controller": "CategoriesController"
      }
    },
...
    "/categories/{id}": {
      "get": {
        "summary": "Get category by ID",
        "description": "Get category by ID",
        "tags": [
          "categories"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "id": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "type": {
                          "oneOf": [
                            {
                              "const": "BASIC"
                            },
                            {
                              "const": "CLINICAL"
                            }
                          ]
                        },
                        "questionCount": {
                          "type": "number"
                        },
                        "isActive": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "name",
                        "id",
                        "displayName",
                        "type",
                        "questionCount",
                        "isActive",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "categories",
          "findById"
        ],
        "x-samchon-controller": "CategoriesController"
      },
      "put": {
        "summary": "Update a category",
        "description": "Update a category",
        "tags": [
          "categories"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "id": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "type": {
                          "oneOf": [
                            {
                              "const": "BASIC"
                            },
                            {
                              "const": "CLINICAL"
                            }
                          ]
                        },
                        "questionCount": {
                          "type": "number"
                        },
                        "isActive": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "name",
                        "id",
                        "displayName",
                        "type",
                        "questionCount",
                        "isActive",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "categories",
          "update"
        ],
        "x-samchon-controller": "CategoriesController"
      },
      "delete": {
        "summary": "Delete a category",
        "description": "Delete a category (only custom categories can be deleted)",
        "tags": [
          "categories"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "content": {
              "application/json": {}
            }
          }
        },
        "x-samchon-accessor": [
          "categories",
          "$delete"
        ],
        "x-samchon-controller": "CategoriesController"
      }
    },
...
  },
  "components": {
    "schemas": {
      "PaginatedResponseDto__type": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "id": {
                  "type": "string"
                },
                "displayName": {
                  "type": "string"
                },
                "type": {
                  "oneOf": [
                    {
                      "const": "BASIC"
                    },
                    {
                      "const": "CLINICAL"
                    }
                  ]
                },
                "questionCount": {
                  "type": "number"
                },
                "isActive": {
                  "type": "boolean"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "name",
                "id",
                "displayName",
                "type",
                "questionCount",
                "isActive",
                "createdAt",
                "updatedAt"
              ]
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "limit": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "totalPages": {
                "type": "number"
              },
              "hasNextPage": {
                "type": "boolean"
              },
              "hasPrevPage": {
                "type": "boolean"
              }
            },
            "required": [
              "page",
              "limit",
              "total",
              "totalPages",
              "hasNextPage",
              "hasPrevPage"
            ]
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "pagination",
          "success"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "type": {
            "oneOf": [
              {
                "const": "BASIC"
              },
              {
                "const": "CLINICAL"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "displayName",
          "type"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "type": {
            "oneOf": [
              {
                "const": "BASIC"
              },
              {
                "const": "CLINICAL"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": []
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}